refactor: moved common rpc hooks to common/rpc folder
