+++ apps/davi/src/components/ProposalInfoCard/types.tsx
@@ -1,4 +1,4 @@
-import { GuildConfigProps } from 'stores/modules/common/fetchers/useGuildConfig';
+import { GuildConfigProps } from 'stores/modules/common/fetchers/rpc/useGuildConfig';
 import { Proposal } from 'types/types.guilds.d';
 
 export interface ProposalInfoCardProps {
