+++ apps/davi/src/components/StakeTokensModal/types.ts
@@ -2,7 +2,7 @@ import { BigNumber } from 'ethers';
 import { ERC20Info } from 'hooks/Guilds/erc20/useERC20Info';
 import { ERC20, ERC20Guild } from 'dxdao-contracts/types';
 import React from 'react';
-import { GuildConfigProps } from 'stores/modules/common/fetchers/useGuildConfig';
+import { GuildConfigProps } from 'stores/modules/common/fetchers/rpc/useGuildConfig';
 export interface StakeTokensModalProps {
   isOpen: boolean;
   onDismiss: () => void;
