+++ apps/davi/src/stores/modules/SnapshotRepGuild/fetchers/rpc/useMemberCount.ts
@@ -1,5 +1,5 @@
 import { ERC20SnapshotRep } from 'contracts/ts-files/ERC20SnapshotRep';
-import { useDAOToken } from 'stores/modules/common/fetchers';
+import { useDAOToken } from 'stores/modules/common/fetchers/rpc';
 import { useContractRead } from 'wagmi';
 
 export const useMemberCount = (daoId: `0x${string}`) => {
