+++ apps/davi/src/stores/modules/SnapshotRepGuild/fetchers/rpc/useTotalLocked.ts
@@ -3,7 +3,7 @@ import { useContractRead } from 'wagmi';
 import { BaseERC20Guild } from 'contracts/ts-files/BaseERC20Guild';
 import { ERC20SnapshotRep } from 'contracts/ts-files/ERC20SnapshotRep';
 import { useListenToTokenTransfer } from '../../events/useListenToTokenTransfer';
-import { useDAOToken, useSnapshotId } from 'stores/modules/common/fetchers';
+import { useDAOToken, useSnapshotId } from 'stores/modules/common/fetchers/rpc';
 import { FetcherHooksInterface } from 'stores/types';
 
 type IUseTotalLocked = FetcherHooksInterface['useTotalLocked'];
