+++ apps/davi/src/stores/modules/common/fetchers/rpc/fixtures.ts
