+++ apps/davi/src/stores/modules/common/fetchers/rpc/useGetPermissions.test.ts
@@ -11,7 +11,7 @@ const mockValueAllowed = BigNumber.from(700000);
 
 jest.mock('contexts/Guilds/orbis', () => ({}));
 
-jest.mock('stores/modules/common/fetchers', () => ({
+jest.mock('stores/modules/common/fetchers/rpc', () => ({
   useProposalCalls: () => ({ options: [] }),
   useGuildConfig: () => ({
     data: {
