+++ apps/davi/src/stores/modules/common/fetchers/rpc/useGetVotes.tsx
