+++ apps/davi/src/stores/modules/common/fetchers/rpc/useGuildProposalIds.ts
@@ -2,7 +2,7 @@ import { BaseERC20Guild } from 'contracts/ts-files/BaseERC20Guild';
 import { useMemo } from 'react';
 import { FetcherHooksInterface } from 'stores/types';
 import { useContractRead } from 'wagmi';
-import { useListenToProposalStateChanged } from '../events/useListenToProposalStateChanged';
+import { useListenToProposalStateChanged } from '../../events/useListenToProposalStateChanged';
 
 type IUseGuildProposalIds = FetcherHooksInterface['useGuildProposalIds'];
