+++ apps/davi/src/stores/modules/common/fetchers/rpc/useIsProposalCreationAllowed.tsx
