+++ apps/davi/src/stores/modules/common/fetchers/rpc/useProposal.ts
