+++ apps/davi/src/utils/fixtures.ts
@@ -10,7 +10,7 @@ import {
   Option,
   SupportedAction,
 } from 'components/ActionsBuilder/types';
-import { GuildConfigProps } from 'stores/modules/common/fetchers/useGuildConfig';
+import { GuildConfigProps } from 'stores/modules/common/fetchers/rpc/useGuildConfig';
 
 export const proposalMock: Proposal = {
   id: '0x1234567890123456789012345678901234567890',
